last

+++ Modules/Scholarship/Database/Migrations/2020_10_16_080458_create_scholarship_academic_achievements.php
@@ -19,7 +19,7 @@ class CreateScholarshipAcademicAchievements extends Migration
             $table->unsignedBigInteger('student_id');
             $table->double('ip');
             $table->tinyInteger('sks');
-            $table->tinyInteger('ipk');
+            $table->double('ipk');
             $table->string('description');
             $table->string('khs');
             $table->timestamps();
